shutdown fsnotify in each test, await receipt of socket messages
if two watchers exist for the same dir, seems like the first one will get events meant for the second one on Ubuntu... So shut down before creating a watcher for the same dir.

Remove sleeps that were in there to wait for socket actions. We can just resolve a promise and be much less flaky.